Finalização das Funções Principais do Sistema
Todas as Questões foram Concluídas
1. Alterações e Adições dos métodos em AlunoDAO
- Adicionado o Sistema de adição de nota em uma disciplina de um aluno.
- Criação do método de checar a existência de uma disciplina em
determinado aluno
- Criação do método de retorno da aprovação em determinada disciplina
2. Todas as Funcionalidades novas e antigas foram testadas na Classe
Principal.

+++ src/com/qst1/vo/Disciplina.java
@@ -3,7 +3,7 @@ package com.qst1.vo;
 public class Disciplina {
 	private String nome;
 	private boolean aprovado;
-	private Float nota;
+	private Double nota;
 	private Integer codigo;
 	
 	private static int geradorCodigo = 0;
@@ -27,11 +27,11 @@ public class Disciplina {
 		this.nome = nome;
 	}
 	
-	public Float getNota(){
+	public Double getNota(){
 		return this.nota;
 	}
 	
-	public void setNota(Float valor){
+	public void setNota(Double valor){
 		this.nota = valor;
 		if(this.nota < 6){
 			this.aprovado = false;
